backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
google-auth
google-api-python-client
google-auth-oauthlib
anthropic
python-jose[cryptography]
python-dotenv
httpx
